Update requirements.txt file
It is customary in Python development to use a file called
`requirements.txt` that contains the list of external modules
that must be installed. When adding a new module to the project
update `requirements.txt` so it's evident how to get the
project up and running for others.

To create/update the file do:

    pip3 freeze > requirements.txt

To install/update packages when you receive an updated
file, do:

    pip3 install -r requirements.txt